requirements.txt
# Python 3.10
numpy
git+https://github.com/fra31/auto-attack.git
spikingjelly
datasets
cupy-cuda12x